+++ packages/plugin-audio/web/src/toolbar/audioPluginService.ts
@@ -7,12 +7,13 @@ export class AudioPluginService implements IMediaPluginService {
     const isCoverImageData = file.type.includes('image');
     if (isCoverImageData) return { componentData };
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    const { html, authorName, coverImage, ...rest } = componentData;
+    const { html, authorName, name, coverImage, ...rest } = componentData;
     return {
       componentData: {
         ...rest,
         audio: { src: { id: url } },
-        name: file.name.split('.')[0],
+        name: name || file.name.split('.')[0],
+        authorName,
       },
       componentState: {
         tempData: true,